requirements.txt
google-adk
python-dotenv
PyPDF2
fastapi
uvicorn